+++ Adapters/IronSource/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## IronSource iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationIronSource', '9.3.0.1'
+pod 'CASMediationIronSource', '9.4.0.0'
 ```
+
+### 9.4.0.0
+- Certified with IronSource - 9.4.0
 
 ### 9.3.0.1
 - Resolved a minor internal configuration conflict.